Performance optimizations: branchless CPU flags, HexUtilities buffer construction, ColorHelper float
- NES CPU: Branchless SetZeroNegativeFlags (PSFlags::Negative=0x80 maps to bit 7)
- SNES CPU: Branchless SetZeroNegativeFlags for both 8-bit and 16-bit variants
- HexUtilities: Direct char buffer construction eliminates string concatenation
  temporaries in ToHex (uint16/20/24/32/64) — up to 2.6x faster for ToHex32
- ColorHelper: double→float conversion (sufficient for 8-bit color channels)
  RgbToHsl allocations: 24KB→12KB per 1024 colors, ~11% faster
- Fix BenchmarkLogicalGroupRule namespace (Configs, not Columns)
- Expand copilot-instructions.md with detailed K&R brace style examples

+++ Tests/Utilities/ColorHelperTests.cs
@@ -102,7 +102,7 @@ public class ColorHelperTests
 	[Fact]
 	public void HslToRgb_RedHsl_ReturnsPureRed()
 	{
-		var hsl = new HslColor { H = 0, S = 1.0, L = 0.5 };
+		var hsl = new HslColor { H = 0, S = 1.0f, L = 0.5f };
 		var color = ColorHelper.HslToRgb(hsl);
 
 		Assert.Equal(255, color.R);
@@ -113,7 +113,7 @@ public class ColorHelperTests
 	[Fact]
 	public void HslToRgb_ZeroSaturation_ReturnsGray()
 	{
-		var hsl = new HslColor { H = 180, S = 0, L = 0.5 };
+		var hsl = new HslColor { H = 180, S = 0, L = 0.5f };
 		var color = ColorHelper.HslToRgb(hsl);
 
 		// With 0 saturation, should be equal R=G=B
@@ -128,9 +128,9 @@ public class ColorHelperTests
 	[InlineData(180)]
 	[InlineData(240)]
 	[InlineData(300)]
-	public void HslToRgb_FullSaturation_HasAtLeastOneMaxComponent(double hue)
+	public void HslToRgb_FullSaturation_HasAtLeastOneMaxComponent(float hue)
 	{
-		var hsl = new HslColor { H = hue, S = 1.0, L = 0.5 };
+		var hsl = new HslColor { H = hue, S = 1.0f, L = 0.5f };
 		var color = ColorHelper.HslToRgb(hsl);
 
 		// At full saturation/half lightness, at least one component is 255
